+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - `RoutingConfiguration.MainLayer` and `RoutingConfiguration.LayerPenalty` are set obsolete.
 - `EdgeInfo.HasVerticalChange` is set obsolete.
 - `AdaptiveGraphRouting.RenderElements` is no longer paint hint lines in two different colors. Instead regular edges are paint into three groups. Weights are included to additional properties of produced elements.
+- Removed rule exception from `AdaptiveGraphRouting` that prevented vertical edges turn cost being discounter.
 
 
 ### Fixed